+++ cmd/rag-eval/cmds/chunk/root.go
@@ -14,6 +14,7 @@ func NewCommand() *cobra.Command {
 
 	cmd.AddCommand(newApplyCommand())
 	cmd.AddCommand(newStrategiesCommand())
+	cmd.AddCommand(newEnrichCommand())
 
 	return cmd
 }